% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{citations}
\alias{citations}
\title{Citation}
\usage{
citations(includeURL = TRUE, bibtex = TRUE)
}
\arguments{
\item{includeURL}{boolean, Should the URL be returned?}

\item{bibtex}{boolean, Should the citations be returned as bibtex?}
}
\value{
printed output to consola
}
\description{
Generates citations of all loaded packages
}
\examples{
library(ggplot2)
library(agricolaeplotr)
library(agricolae)
library(raster)
citations()
}
