% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DOE_obj}
\alias{DOE_obj}
\title{Measures of a Field Design}
\usage{
DOE_obj(p)
}
\arguments{
\item{p}{\code{ggplot} object containing the data of the plot}
}
\value{
a list with several useful information about the experiment and the field
}
\description{
Returns a list with several useful information about the experiment
}
\examples{
library(agricolae)
library(agricolaeplotr)
trt = c(2,3,4,5,6)
outdesign1 <-design.crd(trt,r=5,serie=2,2543,'Mersenne-Twister')
p <- plot_design_crd(outdesign1,
             ncols = 7,
             nrows = 4,
             width = 10,
             height = 10,
             reverse_y = TRUE)
stats <- DOE_obj(p)
stats
}
