% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeler.R
\name{generate_qr_labels}
\alias{generate_qr_labels}
\title{Generate QR Code Labels and Save as PDF}
\usage{
generate_qr_labels(
  input_df,
  output_file = "labelscorrected",
  experiment_name = "My Experiment",
  img_width = 65,
  img_height = 32,
  page_width = 210,
  page_height = 297
)
}
\arguments{
\item{input_df}{A data frame containing label data. The data frame should have at least three columns: Name1, Name2, and Value.}

\item{output_file}{The base name for the generated PDF files (default: "labelscorrected").}

\item{experiment_name}{A string specifying the name of the experiment (default: "My Experiment").}

\item{img_width}{Width of each label in millimeters (default: 70mm).}

\item{img_height}{Height of each label in millimeters (default: 36mm).}

\item{page_width}{Width of the page in millimeters (default: 210mm for A4 size).}

\item{page_height}{Height of the page in millimeters (default: 297mm for A4 size).}
}
\value{
A single PDF file containing all generated QR code labels, arranged in a grid layout.
}
\description{
This function generates QR code labels based on the provided dataset and saves them as a PDF file.
Each label contains a QR code along with associated metadata. The labels are arranged in a grid on an A4-sized page.
}
\details{
- The function calculates the number of labels that fit on an A4 page.
- Labels are distributed in a grid layout.
- Each label includes a QR code and text extracted from the input data.
- Multiple pages are handled automatically, and all generated PDFs are merged into a single output file.
}
\examples{
library(agricolaeplotr)
library(agricolae)
library(qrcode)

trt <- c(2, 4)
k <- 6
outdesign <- design.ab(trt, r = k, serie = 3, design = 'rcbd')
generate_qr_labels(outdesign$book)

# Triple lattice design (9 treatments)
trt <- LETTERS[1:9]
outdesign <- design.lattice(trt, r = 3, serie = 2)
generate_qr_labels(outdesign$book)

t1 <- c("A", "B", "C", "D", "E")
t2 <- c(1, 2, 3, 4, 5, 6)
outdesign <- design.split(t1, t2, r = 3, serie = 2,
 seed = 45, kinds = "Super-Duper") # Split-plot design
generate_qr_labels(outdesign$book, output_file = "mylabels2")
}
