% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{make_polygons_ggvoronoi}
\alias{make_polygons_ggvoronoi}
\title{Extract and Transform Voronoi Polygons}
\usage{
make_polygons_ggvoronoi(
  ggplot_object,
  projection_input = 25832,
  projection_output = 4326
)
}
\arguments{
\item{ggplot_object}{A ggplot object containing Voronoi polygons generated using ggvoronoi package.}

\item{projection_input}{The input coordinate reference system (CRS) used in the Voronoi polygons.
Default value: 25832 (EPSG:25832, UTM Zone 32N - Europe)}

\item{projection_output}{The desired output coordinate reference system (CRS) for the transformed Voronoi polygons.
Default value: 4326 (EPSG:4326, WGS 84 - Latitude/Longitude)}
}
\value{
A SpatialPolygonsDataFrame object containing the transformed Voronoi polygons.
}
\description{
This function extracts Voronoi polygons from a ggplot object created using the ggvoronoi package
and transforms their coordinate reference system (CRS).
}
\details{
This function extracts the Voronoi polygons from the input ggplot object and converts them into a SpatialPolygonsDataFrame.
The function assumes that the ggplot object was created using the ggvoronoi package with stat_voronoi(geom="path") and geom_point().
Make sure to load the required libraries (ggplot2, ggvoronoi, sp) before using this function.
}
\examples{
\dontrun{
library(ggplot2)
library(ggvoronoi)
library(sf)

df <- structure(list(groups = structure(1:16, .Label = c("0", "1",
"2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13",
"14", "15"), class = "factor"),
latitude = c(454958.780465603,
455004.47280591, 454885.338464729, 454891.038200793, 454911.54140957,
454947.515894516, 454845.494336779, 454950.428690276, 455001.684028287,
454936.964919935, 455017.721815154, 455084.876168519, 454884.307056065,
455053.71370209, 454988.433646113, 454933.070817933),
longitude = c(5936577.41802353,
5936650.33357495, 5936684.00492433, 5936603.3723963, 5936644.3710608,
5936678.18028261, 5936661.25833358, 5936614.7618022, 5936693.37183672,
5936727.14434355, 5936609.08891642, 5936620.03577825, 5936727.23617373,
5936663.54803475, 5936737.77699043, 5936769.89451942)),
row.names = c(NA,-16L), class = "data.frame")

p <- ggplot(df,aes(latitude,longitude)) +
     stat_voronoi(geom="path") +
     geom_point()
p
polygons_list <- make_polygons_ggvoronoi(p,
 projection_input = 25832,
 projection_output = 4326)

plot(polygons_list)
}
}
\seealso{
\code{\link{stat_voronoi}}
\code{\link{geom_point}}
}
\keyword{spatial}
