% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_latin_square}
\alias{plot_latin_square}
\title{Plot Latin Square Design}
\usage{
plot_latin_square(
  design,
  x = "col",
  y = "row",
  factor_name = "trt",
  labels = "plots",
  width = 1,
  height = 1,
  space_width = 0.95,
  space_height = 0.85,
  reverse_y = FALSE,
  reverse_x = FALSE
)
}
\arguments{
\item{design}{outdesign from \code{agricolae} package}

\item{x}{Describes the x coordinates of a experiment design}

\item{y}{Describes the y coordinates of a experiment design}

\item{factor_name}{Which factor should be used for plotting, needs to be a column in outdesign$book}

\item{labels}{Describes the column from that the plots are taken to display them}

\item{width}{numeric value, describes the width of a plot in an experiment}

\item{height}{numeric value, describes the height of a plot in an experiment}

\item{space_width}{numeric value, describes the share of the space of the plots. 0=only space, 1=no space between plots in term of width}

\item{space_height}{numeric value, describes the share of the space of the plots. 0=only space, 1=no space between plots in term of height}

\item{reverse_y}{boolean, should the plots of the experiment be changed in reverse order in Row direction? use reverse_y=TRUE to have same sketch as in agricolae. default:reverse_y=FALSE}

\item{reverse_x}{boolean, should the plots of the experiment be changed in reverse order in column direction? default:reverse_x=FALSE}
}
\value{
\code{ggplot} graphic that can be modified, if wished
}
\description{
Plot a design of a factorial  experiment with a latin square design from \code{agricolae} design.lsd
}
\examples{
library(agricolaeplotr)
library(agricolae)
trt<-LETTERS[1:9]
outdesign<- design.lsd(trt,serie=2)
plot_latin_square(outdesign, reverse_y = TRUE)
}
