% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{serpentine}
\alias{serpentine}
\title{Serpentine}
\usage{
serpentine(n, times, m = 1)
}
\arguments{
\item{n}{integer value indicating the upper cap of a numeric sequence}

\item{times}{integer number of replications}

\item{m}{integer value indicating the lower
cap of a numeric sequence}
}
\value{
vector containing the serpentine sequence
}
\description{
This function produces a serpentine array of integers beginning by one
}
\examples{
serpentine(n=20,times = 15)
serpentine(n=20,times = 15,m=4)
}
