% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_input_extend}
\alias{test_input_extend}
\title{Test if input for width and height is numeric}
\usage{
test_input_extend(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
Test if input is numeric for field width and height
}
\examples{
library(agricolaeplotr)
test_input_extend(3)
}
