% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_input_height}
\alias{test_input_height}
\title{Test if input for height is numeric}
\usage{
test_input_height(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
Test if input is numeric for field height
}
\examples{
library(agricolaeplotr)
test_input_height(5)
}
