% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_input_ncols}
\alias{test_input_ncols}
\title{checks matrix column input}
\usage{
test_input_ncols(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
checks if input is suitable for matrix column indication
}
\examples{
library(agricolaeplotr)
test_input_ncols(9)
}
