% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_input_nrows}
\alias{test_input_nrows}
\title{checks matrix rows input}
\usage{
test_input_nrows(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
checks if input is suitable for matrix row indication
}
\examples{
library(agricolaeplotr)
test_input_nrows(10)
}
