% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_input_reverse}
\alias{test_input_reverse}
\title{Test if input is a logical}
\usage{
test_input_reverse(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
Test if input is a logical
}
\examples{
library(agricolaeplotr)
test_input_reverse(TRUE)
}
