% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_input_shift}
\alias{test_input_shift}
\title{Test if input for shift parameter is numeric}
\usage{
test_input_shift(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
Test if input is numeric for shift parameter
}
\examples{
library(agricolaeplotr)
test_input_shift(0.5)
}
