% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_input_width}
\alias{test_input_width}
\title{Test if input for width is numeric}
\usage{
test_input_width(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
Test if input is numeric for field width
}
\examples{
library(agricolaeplotr)
test_input_width(3)
}
