% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_name_in_column}
\alias{test_name_in_column}
\title{Test if input column names}
\usage{
test_name_in_column(x, design)
}
\arguments{
\item{x}{string input}

\item{design}{design from \code{agricolae} package}
}
\value{
error
}
\description{
Test if input is in column names of a table
}
\examples{
library(agricolaeplotr)
library(agricolae)
trt<-c(2,4)
k=6
outdesign<-design.ab(trt, r=k, serie=3,design='rcbd')
test_name_in_column('B',outdesign)
}
