% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_names_design}
\alias{test_names_design}
\title{Test of experimental design}
\usage{
test_names_design(design)
}
\arguments{
\item{design}{design from \code{agricolae} package}
}
\value{
error
}
\description{
Test if the outdesign file contains book and parameter list
}
\examples{
library(agricolaeplotr)
library(agricolae)
trt<-c(2,4)
k=6
outdesign<-design.ab(trt, r=k, serie=3,design='rcbd')
test_names_design(outdesign)
}
