% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_string}
\alias{test_string}
\title{Test if input is a string}
\usage{
test_string(x)
}
\arguments{
\item{x}{input to be tested}
}
\value{
error
}
\description{
Test if input is a string
}
\examples{
library(agricolaeplotr)
test_string('smallstring')
}
