% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{theme_gi}
\alias{theme_gi}
\title{theme_gi}
\usage{
theme_gi()
}
\value{
a 'ggplot' graph with a modified theme
}
\description{
Creates a theme for 'ggplot' based graphics to ensure
to meet formal requirements for conferences of the
Gesellschaft fuer Informatik
}
\examples{
# example borrowed from ggplot2
library(ggplot2)
df <- data.frame(
gp = factor(rep(letters[1:3], each = 10)),
y = rnorm(30))

p <- ggplot() +
geom_point(data = df, aes(gp, y))
p <- p + theme_gi();p
}
