% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{theme_gil}
\alias{theme_gil}
\title{theme_gil}
\usage{
theme_gil()
}
\value{
a 'ggplot' graph with a modified theme
}
\description{
Creates a theme for 'ggplot' based graphics to ensure
to meet formal requirements for conferences of the
Gesellschaft fuer Informatik in der Land-, Forst,- und Ern?hrungswirtschaft e.V. (GIL).
}
