% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{theme_pres}
\alias{theme_pres}
\title{ggplot2 theme for outdoor presentation}
\usage{
theme_pres()
}
\value{
ggplot2 theme
}
\description{
This theme is designed to increase font size to ensure readability on outdoor used devices
}
\examples{
library(agricolaeplotr)
library(agricolae)
T1<-c('a','b','c','d','e','f','g')
T2<-c('v','w','x','y','z')
r <- 4
outdesign2 <- design.split(trt1=T1, trt2=T2, r=r,
serie = 2, seed = 0, kinds = 'Super-Duper',
randomization=FALSE,first=TRUE,design = 'crd')
plot_split_crd(outdesign2,ncols = 6,nrows=5)+
theme_pres()
}
