% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_fieldhub}
\alias{plot_fieldhub}
\title{Plot FielDHub Design}
\usage{
plot_fieldhub(
  design,
  x = "COLUMN",
  y = "ROW",
  labels = "PLOT",
  factor_name = "TREATMENT",
  width = 1,
  height = 1,
  space_width = 0.95,
  space_height = 0.85,
  reverse_y = FALSE,
  reverse_x = FALSE
)
}
\arguments{
\item{design}{outdesign from \code{FielDHub} package with on of the following IDs: c(9,13,14,15,16)}

\item{x}{Describes the x coordinates of a experiment design}

\item{y}{Describes the y coordinates of a experiment design}

\item{labels}{string Describes the column from that the plots are taken to display them}

\item{factor_name}{string Which factor should be used for plotting, needs to be a column in outdesign$book}

\item{width}{numeric value, describes the width of a plot in an experiment}

\item{height}{numeric value, describes the height of a plot in an experiment}

\item{space_width}{numeric value, describes the share of the space of the plots. 0=only space, 1=no space between plots in term of width}

\item{space_height}{numeric value, describes the share of the space of the plots. 0=only space, 1=no space between plots in term of height}

\item{reverse_y}{boolean, should the plots of the experiment be changed in reverse order in Row direction? use reverse_y=TRUE to have same sketch as in agricolae. default:reverse_y=FALSE}

\item{reverse_x}{boolean, should the plots of the experiment be changed in reverse order in column direction? default:reverse_x=FALSE}
}
\value{
\code{ggplot} graphic that can be modified, if wished
}
\description{
Plots designs from \code{FielDHub} package
}
\examples{
library(agricolaeplotr)
library(FielDHub)
SpatpREP1 <- partially_replicated(nrows = 25,
                                  ncols = 18,
                                  repGens = c(280,50,10,1,1),
                                  repUnits = c(1,2,3,20,20),
                                  planter = "cartesian",
                                  plotNumber = 101,
                                  seed = 77)
SpatpREP1$infoDesign

plot_fieldhub(SpatpREP1,
labels = "PLOT",
factor_name = "PLOT",
width = 12,
height = 10,
reverse_y = TRUE,
reverse_x = TRUE)

NAME <- paste("G", 1:492, sep = "")
repGens = c(108, 384);repUnits = c(2,1)
REPS <- rep(repUnits, repGens)
treatment_list <- data.frame(list(ENTRY = 1:492, NAME = NAME, REPS = REPS))

SpatpREP2 <- partially_replicated(nrows = 30,
                                  ncols = 20,
                                  planter = "serpentine",
                                  plotNumber = 101,
                                  seed = 41,
                                  data = treatment_list)
SpatpREP2$infoDesign

plot_fieldhub(SpatpREP2,
labels = "PLOT",
factor_name = "PLOT",
width = 12,
height = 10,
reverse_y = TRUE,
reverse_x = TRUE)

}
